Move SourceOptions into infra.passes.options
Just a simple code organization change consistent with #2018 and the
ongoinig work on #1174.
