--- mod-infra/src/main/scala/at/forsyte/apalache/infra/passes/SourceOption.scala
@@ -1,20 +1,0 @@
-package at.forsyte.apalache.infra.passes
-
-/** Defines the data sources supported in the [[PassOptions]] */
-sealed abstract class SourceOption
-
-object SourceOption {
-
-  /** Data to be loaded from a file */
-  final case class FileSource(file: java.io.File) extends SourceOption
-
-  /**
-   * Data supplied as a string
-   *
-   * @param content
-   *   the principle data source
-   * @param aux
-   *   auxiliary data sources
-   */
-  final case class StringSource(content: String, aux: Seq[String] = Seq()) extends SourceOption
-}
