+++ mod-infra/src/main/scala/at/forsyte/apalache/infra/passes/options.scala
@@ -3,6 +3,25 @@ package at.forsyte.apalache.infra.passes
 /** Collects the passoptions supported for configuring Apalache's various passes */
 object options {
 
+  /** Defines the data sources supported in the [[PassOptions]] */
+  sealed abstract class SourceOption
+
+  object SourceOption {
+
+    /** Data to be loaded from a file */
+    final case class FileSource(file: java.io.File) extends SourceOption
+
+    /**
+     * Data supplied as a string
+     *
+     * @param content
+     *   the principle data source
+     * @param aux
+     *   auxiliary data sources
+     */
+    final case class StringSource(content: String, aux: Seq[String] = Seq()) extends SourceOption
+  }
+
   /**
    * Defines the SMT encoding options supported
    *
